+++ src/pararnn/cells/__init__.py
@@ -4,6 +4,7 @@ from pararnn.cells.para_hopfield import ParaHopfield
 from pararnn.cells.para_lstm import ParaLSTM
 from pararnn.cells.para_m2rnn import ParaM2RNN
 from pararnn.cells.para_nlru import ParaNLRU
+from pararnn.cells.para_rwkv7 import ParaRWKV7
 from pararnn.cells.para_slstm import ParaSLSTM
 from pararnn.cells.protocol import RNNCell, check_cell
 
@@ -14,6 +15,7 @@ __all__ = [
     "ParaLSTM",
     "ParaM2RNN",
     "ParaNLRU",
+    "ParaRWKV7",
     "ParaSLSTM",
     "RNNCell",
     "check_cell",